+++ packages/features/src/budget/BudgetFormContent/index.ts
@@ -1,0 +1,6 @@
+export { BudgetFormContent } from './BudgetFormContent.native'
+export type {
+  IBudgetFormContentProps,
+  IBudgetFormRef,
+  IBudgetFormValues,
+} from './BudgetFormContent'