+++ packages/features/src/budget/BudgetFormContent/index.web.ts
@@ -1,0 +1,6 @@
+export { BudgetFormContent } from './BudgetFormContent.web'
+export type {
+  IBudgetFormContentProps,
+  IBudgetFormRef,
+  IBudgetFormValues,
+} from './BudgetFormContent'