+++ packages/features/src/modal/index.ts
@@ -1,2 +1,0 @@
-export { ModalProvider } from './AppModal.native'
-export type { IModalProviderProps } from './AppModal'