+++ packages/features/src/modal/index.web.ts
@@ -1,2 +1,0 @@
-export { ModalProvider } from './AppModal.web'
-export type { IModalProviderProps } from './AppModal'