+++ packages/features/src/shared/ModalRenderer/index.ts
@@ -1,0 +1,2 @@
+export { ModalRenderer } from './ModalRenderer.native'
+export type { IModalRendererProps } from './ModalRenderer'