+++ packages/features/src/shared/ModalRenderer/index.web.ts
@@ -1,0 +1,2 @@
+export { ModalRenderer } from './ModalRenderer.web'
+export type { IModalRendererProps } from './ModalRenderer'